Bug fix for the Needleman-Wunsch sequence alignment algorithm.
Part of the scoring system was functioning incorrectly when the gap penalty scores were non-integer,
as some scores were being stored in an integer array.  Now the array is a float array.

+++ lib/sequence_alignment/needleman_wunsch.py
@@ -33,7 +33,7 @@ from lib.errors import RelaxError, RelaxFault
 SCORE_MATCH = 1
 SCORE_MISMATCH = -1
 SCORE_GAP_PENALTY = 1
-SCORES = zeros(3, int16)
+SCORES = zeros(3, float32)
 
 # Indices.
 TRACEBACK_DIAG = 0
